+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- On Windows, the desktop app now uses a custom title bar.
+- On Windows, the desktop app now uses a custom title bar, and a custom menu bar powered by [OS-GUI.js](https://os-gui.js.org)
 - The background of the whole window is now light purple instead of white or black, matching the background color of main part of the app.
 
 ### Fixed